added new lib for float CW
